Do not kill dnsmasq instances started by init.
When specifying a specific DNS server in /etc/network/interfaces:

auto eth0
iface eth0 inet dhcp
    dns-nameservers 8.8.8.8

and having the same dns-server specified in /etc/dhcp/dhclient.conf:

prepend domain-name-servers 8.8.8.8;

and restarting the machine will make stack.sh to fail because the
dnsmasq created by init is deleted, consequently, it won't be possible
to download the cirros image.

With this patch, stack.sh won't kill dnsmasq instances sarted by
/sbin/init (pid 1).

Closes-Bug: #1341885

+++ stack.sh
@@ -1040,12 +1040,13 @@ fi
 
 if is_service_enabled n-net q-dhcp; then
     # Delete traces of nova networks from prior runs
-    # Do not kill any dnsmasq instance spawned by NetworkManager
+    # Do not kill any dnsmasq instance spawned by NetworkManager or /sbin/init.
     netman_pid=$(pidof NetworkManager || true)
-    if [ -z "$netman_pid" ]; then
+    init_pid=$(pidof init || true)
+    if [ -z "$netman_pid" -a -z "$init_pid"]; then
         sudo killall dnsmasq || true
     else
-        sudo ps h -o pid,ppid -C dnsmasq | grep -v $netman_pid | awk '{print $1}' | sudo xargs kill || true
+        sudo ps h -o pid,ppid -C dnsmasq | grep -v $netman_pid | grep -v $init_pid | awk '{print $1}' | sudo xargs kill || true
     fi
 
     clean_iptables
